testの結果radix_sortが間違えている

+++ include/push_swap.h
@@ -46,7 +46,7 @@ void	sort_size4(t_list **stack_a, t_list **stack_b);
 void	sort_size5(t_list **stack_a, t_list **stack_b);
 void	simple_sort(t_list **stack_a, t_list **stack_b);
 void	radix_sort(t_list **stack_a, t_list **stack_b);
-int		count_max_bits(t_list **stack);
+size_t	count_max_bits(t_list **stack);
 void	free_list(t_list *list);
 void	check_args(char **args, int start);
 void	put_error(char *str);
